refac: use double precision for raw material quantities + add method to reset product's global id correctly

+++ src/main/java/domain/entities/product/CopoDeVidro.java
@@ -2,7 +2,7 @@ package domain.entities.product;
 
 public class CopoDeVidro extends Product {
 
-    public CopoDeVidro(String name, int rawMaterialPerUnit) {
+    public CopoDeVidro(String name, double rawMaterialPerUnit) {
         super(name, rawMaterialPerUnit, 0.4, 0.0);
     }
